IoT_master/requirements.txt
cryptography
opencv-python
parameterized>=0.7
pyyaml
requests
